requirements.txt
discord.py
flask
waitress
gtts
ffmpeg
dblpy
pynacl
